requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.0.1
tenacity==8.2.3